+++ .vscode/settings.json
@@ -1,0 +1,28 @@
+{
+    "cSpell.words": [
+        "Augustana",
+        "Brkln",
+        "colwidth",
+        "Concordia",
+        "CUSA",
+        "downcasting",
+        "Embry",
+        "fastbbreak",
+        "foulon",
+        "freethrow",
+        "fromturnover",
+        "isna",
+        "jumpball",
+        "MAAC",
+        "Marymount",
+        "Maryville",
+        "MEAC",
+        "offensivedeadball",
+        "pointsinthepaint",
+        "startperiod",
+        "SWAC",
+        "totalcol",
+        "Trpl",
+        "Wheaton"
+    ]
+}
