+++ CHANGELOG.md
@@ -1,0 +1,8 @@
+# NCAApy Changelog
+
+## 0.0.1:
+- Refactored project into a Python package
+- Refactored the code to comply with PEP formatting rules.
+- Cleaned up the code so all imports are explicit, instead of implicit.
+- Added `/.github/workflows/python-publish.yml` to enable PyPi package publishing for this project.
+- Set the package version to `0.0.1`.
