--- variables.py
@@ -1,101 +1,0 @@
-
-years = {
-    2024: 18221,
-    2023: 16060,
-    2022: 15881,
-    2021: 15480,
-    2020: 15061,
-    2019: 14300
-}
-
-
-conferences = {
-    'AAC': 823,
-    'ACC': 821,
-    'ASun': 920,
-    'AE': 845,
-    'A10': 820,
-    'B12': 25354,
-    'BE': 30184,
-    'BSky': 825,
-    'BSouth': 826,
-    'B10': 827,
-    'BWest': 904,
-    'CAA': 837,
-    'CUSA': 24312,
-    'Horizon': 881,
-    'Ivy': 865,
-    'MAAC': 871,
-    'MAC': 875,
-    'MEAC': 876,
-    'MVC': 884,
-    'MW': 5486,
-    'NEC': 5486,
-    'OVC': 902,
-    'P12': 905,
-    'Patriot': 838,
-    'SEC': 911,
-    'SWAC': 916,
-    'SoCon': 912,
-    'Southland': 914,
-    'Summit': 819,
-    'Sun Belt': 818,
-    'Fun Belt': 818,
-    'WAC': 923,
-    'WCC': 922
-}
-
-tournaments = {
-    'NCAA': 24062,
-    'NIT': 24084,
-    'Ivy': 23925,
-    'AAC': 23402,
-    'MAC': 23782,
-    'A10': 23863,
-    'B12': 23884,
-    'BE': 23885,
-    'BSky': 23923,
-    'B10': 23942,
-    'MEAC': 23762,
-    'MW': 23847,
-    'P12': 23924,
-    'SEC': 23926,
-    'SWAC': 23862,
-    'Southland': 23922,
-    'ACC': 23566,
-    'BWest': 23742,
-    'CUSA': 23927,
-    'MAAC': 23743,
-    'WAC': 23902,
-    'AE': 23344,
-    'CAA': 23662,
-    'Summit': 23562,
-    'SoCon': 23565,
-    'MVC': 23622,
-    'Sun Belt': 23564,
-    'Fun  Belt': 23564,
-    'WCC': 23642,
-    'BSouth': 23502,
-    'OVC': 23542,
-    'ASun': 23542,
-    'Horizon': 23522,
-    'Patriot': 23362,
-    'NEC': 23543
-}
-
-weird_names = {'Trinity (CT)', 'Anderson (IN)', "St. Mary's (TX)", 'Centenary (NJ)', 'North Central (MN)',
-               "St. Mary's (MD)", 'Notre Dame (OH)', 'Dominican (NY)', "Saint John's (MN)", 'Queens (NC)',
-               'Wayne St. (MI)', 'Bridgewater (VA)', 'Saint Joseph (CT)', 'Charleston (WV)', 'Mt. St. Mary (NY)',
-               'Anderson (SC)', 'Central St. (OH)', 'Salem (WV)', "Saint Mary's (CA)", 'St. Thomas (MN)',
-               'Albany St. (GA)', "St. Joseph's (Brkln)", 'Augustana (IL)', 'Monmouth (IL)', 'Emmanuel (GA)',
-               'Rochester (NY)', 'Bethany (WV)', 'Miami (OH)', 'Marian (WI)', 'Centenary (LA)', 'Notre Dame (MD)',
-               'Southwestern (TX)', 'Westminster (PA)', 'Embry-Riddle (FL)', 'Regis (MA)', 'Carroll (WI)',
-               'Trinity (TX)', 'Bethel (MN)', 'Thomas (ME)', 'Marymount (VA)', 'JWU (Providence)', 'Westminster (UT)',
-               'LMU (CA)', 'Augustana (SD)', 'Ozarks (AR)', 'Wayne St. (NE)', 'King (TN)', 'Clark (MA)',
-               "Saint Mary's (MN)", 'Dominican (IL)', 'Concordia (TX)', 'York (NY)', 'Indiana (PA)', 'Union (TN)',
-               'Pacific (OR)', 'Maryville (MO)', 'Lincoln (MO)', 'Maryville (TN)', 'Wilmington (DE)', 'Regis (CO)',
-               'Central (IA)', 'Westminster (MO)', 'California (PA)', 'York (PA)', 'Benedictine (IL)', 'Miami (FL)',
-               'Wheaton (IL)', 'North Central (IL)', 'Dominican (CA)', 'Lincoln (PA)', 'Queens (NY)',
-               'Saint Francis (PA)', 'Wesleyan (CT)', "St. John's (NY)", 'Wheaton (MA)', 'Emmanuel (MA)', 'Union (NY)',
-               "St. Joseph's (L.I.)", 'St. Thomas (TX)', 'Wilmington (OH)', 'Crown (MN)', "King's (PA)",
-               "St. Joseph's (ME)"}
